Remove some TODOs by reworking/correcting how coordinators/recovery coordinators issue BlockOnDeps
BlockOnDeps can end up getting an insufficient response (not 100% clear on the full range of cases where this can happen). I think it can happen if an earlier part of the transaction is missed by a node like Preaccept so it never got the transaction description?

So modify ApplyAndWaitUntilApplied to always set up the listener even if the Apply fails and respond with not committed.

Then the coordinator (recovery or otherwise) now responds with a maximal Apply (not commit which was a mistake) to enable the application
to proceed.